Fixed the galloping join jumping into matching blocks.
